ColorSlider.cs fix: remove dubious constraints from min/max property setters
Accepting or rejecting property values should not (unnecessarily)
depend on the order in which properties are set. Make Minimum and
Maximum setters symmetric (i.e. acting in the same way): There should
not be one of them throwing in a situation where the other one don't.

An apparently innocent setup done by code like
  Minimum = 200; Maximum = 300;
did throw. This one will not occur in code generated by Designer where
properties happen to be assigned in alphabetic order; yet a generated
  Maximum = -100; Minimum = -200;
suffered from the same problem: setting a negative value on Maximum
failed _silently_ as long as Minimum comes with an initial value of zero.

Change code in such way that assigning reasonable limits (i.e. max >= min)
will succeed in any order and regardless of the previous state; for invalid
parameters (min > max) the assignment occurring later takes precedence
(and pushes the other bound away).

Note that the Value property is (and will be) intentionally constrained
by Minimum and Maximum, and thus is subject to similar restrictions
on the initialization order. This is less of an issue only because
the property name "Value" comes near the end in alphabetic order.