ColorSlider.cs: remove pointless constraint from ScaleSubDivisions property
Originally meant to ensure that minor scale intervals will not (nominally)
be smaller than the resolution given by the underlying integer type; now
obsolete as we can have fractional slider values as well as sub-integer
tick mark increments.

Pointless anyway because the expression
  (_maximum - _minimum) / (_scaleSubDivisions * _scaleDivisions) > 0
on rational numbers will always be true as long as both scaleDivisions
properties are positive (which is checked immediately before said expression,
and is enforced elsewhere), and as long as maximum is greater than minimum
(in other words: unless both limits coincide).
(When needed, the original meaning might be restored by changing >0 into >=1.)